chromium/third_party/blink/web_tests/wpt_internal/origin-isolation/otoi-old-getter-exists.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>The legacy originIsolationRestricted getter still exists for now</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<!--
We're keeping originIsolationRestricted as an alias to originIsolated for
the duration of the origin trial, to avoid breaking any code that depends on it.
This test can be deleted when we finish off https://crbug.com/1103866 by
removing the alias.
-->

<script type="module">
test(() => {
  assert_equals(window.originIsolationRestricted, window.originIsolated,
    "legacy getter value must equal the new getter value");
  assert_equals(window.originIsolationRestricted, true,
    "the getters must return true for this very simple isolation case");
});
</script>
